Update ThreeColorMask.gdshader
Change the order of color_b and color_g to match the color order of rgb instead of rbg. Godot display vec3 (for brightness) as x (red), y (green), z (blue) and converting it to color would make it harder to read. I could split brightness into three floats, but I find it easier to read as a vector (for now).